refactor(generator): 拆分共享 CSS 生成管线
